% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexMap.R
\name{flexMAP}
\alias{flexMAP}
\title{flexMAP: Flexible Modelling of Alternative Polyadenylation}
\usage{
flexMAP(InputTable, path = NULL, covariates = NULL, polyAsite,
  reference, TINfilter = NULL, num_samples = 1, link = "logit",
  type = "ML", link.phi = "log", name = "flexMAP_out",
  normalise = FALSE, exprFilt = 20)
}
\arguments{
\item{InputTable}{Dataframe containing names of bam files in the first column, condition in format 1,2,3..(and so on) in the second column. Additional columns with optional covariates for modelling with names of covariates as column names.}

\item{path}{Path to folder containg indexed bam files}

\item{covariates}{names of covariates (additional to condition, the same as column names in InputTable)}

\item{polyAsite}{Dataframe containing polyadenylation sites in format: GeneID, chromosome, strand, genomic coordinate}

\item{reference}{Dataframe containing reference in format: GeneID, transID, chromosome, strand, transcript start, transcript end, CDS start, CDS end}

\item{TINfilter}{optional, a vector of gene symbols to be removed from analysis because of degradation bias. The TIN values can be obtained by running RSeQC software. The genes with low TIN values have high 3' bias and should be removed from analysis}

\item{num_samples}{minimal number of samples with ratio (0,1)}

\item{link}{default "logit", options: "logit", "probit", "cloglog", "cauchit", "log", "loglog", see ?betareg for explanation}

\item{type}{default "ML", options: "ML", "BC", "BR", see ?betareg for explanation}

\item{link.phi}{default "log", options:  "identity", "log", "sqrt",  see ?betareg for explanation}

\item{name}{optional name of output file if different from default}

\item{normalise}{Default FALSE. Optionally the counts can be normalised for library sizes}

\item{exprFilt}{minimal number counts in short region in all samples, default 20}
}
\value{
An object containing p-values, adjusted p-values for each tested polyadenylation proximal site of a gene. In addition the output contain ratios for each sample. THe output files is also written out to working directory.
}
\description{
A beta-regression-based method for discovering differential alternative polyadenylation events in standard RNA-seq data.
}
\examples{
#Format of InputTable dataframe
head(InputTable)
Sample         Condition Sex
sample1.bam    1         1
sample2.bam    1         2
sample3.bam    2         1
sample4.bam    2         2

#Format of reference dataframe
head(reference)
GeneID  TransID   Chromosome  Strand  TransStart  TransEnd  CDSstart  CDSend
BHLHE23 NM_080606 chr20       -       61637330    61638387  61637400  61638126
CRYBA2  NM_057094 chr2        -       219854911   219858127 219854973 219857898
FNDC8   NM_017559 chr17       +       33448630    33457751  33448712  33457453

#Format of polyAsite dataframe
head(polyAsite)
GeneID Chromosome Strand GenomicCoordinate
CYP2R1 chr11      +      14910363
LSM1   chr8       +      38021091
SIK3   chr11      -      116820211

#Highly degraded genes with 3'bias shoud be removed from analysis
TINfilter = c(CYP2R1, LSM1)

#Run Analysis
flexMAP(InputTable, path='/home/bamFolder/', covariates=c('sex'), polyAsite, reference, TINfilter = c(CYP2R1, LSM1), num_samples = 1, link ='logit', type = 'ML', link.phi='log', name = 'flexMAP_out' , normalise=FALSE, exprFilt = 20)




}
